src/back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aerich (>=0.9.2,<0.10.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "tomlkit (>=0.13.3,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "config_reader.TORTOISE_ORM"
location = "./db/migrations"
src_folder = "./."
